biosphere: Better component grouping
MeshTreeElement: Don't inherit from SceneElement, simplify interface

Tests: Update MeshTree tests + improve parametrisation

MeshTree: Remove kernel_dict()
