+++ eradiate/scenes/biosphere/__init__.py
@@ -1,16 +1,16 @@
-from ._canopy_element import CanopyElement
-from ._core import BiosphereFactory, Canopy
-from ._discrete import DiscreteCanopy, InstancedCanopyElement
+from ._core import BiosphereFactory, Canopy, CanopyElement, InstancedCanopyElement
+from ._discrete import DiscreteCanopy
 from ._leaf_cloud import LeafCloud
-from ._mesh_tree_element import MeshTreeElement
-from ._tree import AbstractTree, MeshTree
+from ._tree import AbstractTree, MeshTree, MeshTreeElement
 
 __all__ = [
+    "AbstractTree",
     "BiosphereFactory",
     "Canopy",
+    "CanopyElement",
     "DiscreteCanopy",
     "InstancedCanopyElement",
     "LeafCloud",
-    "AbstractTree",
-    "CanopyElement",
+    "MeshTree",
+    "MeshTreeElement",
 ]
